lambda/rss-crawler/requirements.txt
feedparser
python-dateutil